# services/svc-payment/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Parking - Payment Service API",
        "description": "OpenAPI docs cho svc-payment",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8002/api",
            "description": "Local API through svc-auth"
        }
    ],
    "paths": {
...
        "/parking-lots/{id}/slots": {
            "get": {
                "tags": [
                    "🏢 Parking Lots"
                ],
                "summary": "Lấy danh sách chỗ đỗ trong một bãi cụ thể",
                "description": "Trả về danh sách tất cả slots trong bãi với thông tin chi tiết",
                "operationId": "7ad86962d1b9d22af0caf9679f941f87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bãi đỗ xe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách chỗ đỗ trong bãi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ParkingSlot"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy bãi xe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Không tìm thấy bãi xe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/parking-lots/{id}/statistics": {
            "get": {
                "tags": [
                    "🏢 Parking Lots"
                ],
                "summary": "Lấy thống kê chi tiết của bãi đỗ xe",
                "description": "Trả về thống kê tổng quan và theo loại xe của bãi đỗ",
                "operationId": "060ab1c407a055629b48502f4e113218",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của bãi đỗ xe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thống kê chi tiết bãi đỗ xe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParkingLotStatistics"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy bãi xe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Không tìm thấy bãi xe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/parking-lots/{id}/stream": {
            "get": {
                "tags": [
                    "🏢 Parking Lots"
                ],
                "summary": "Stream trạng thái slot theo thời gian thực",
                "description": "Server-Sent Events để cập nhật trạng thái slot liên tục",
                "operationId": "71318d34ccbac7a19905d0a921d4854c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bãi đỗ xe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream",
                        "content": {
                            "text/event-stream": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ParkingLotStatistics": {
                "title": "Parking Lot Statistics",
                "description": "Thống kê chi tiết bãi đỗ xe",
                "properties": {
                    "parking_lot_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parking_lot_name": {
                        "type": "string",
                        "example": "Tầng hầm B1"
                    },
                    "summary": {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 100
                            },
                            "available": {
                                "type": "integer",
                                "example": 45
                            },
                            "hold": {
                                "type": "integer",
                                "example": 15
                            },
                            "occupied": {
                                "type": "integer",
                                "example": 40
                            },
                            "utilization_rate": {
                                "type": "number",
                                "example": 55
                            }
                        },
                        "type": "object"
                    },
                    "by_vehicle_type": {
                        "properties": {
                            "motorbike": {
                                "properties": {
                                    "total": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "available": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "hold": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "occupied": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            },
                            "car_4_seat": {
                                "properties": {
                                    "total": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "available": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "hold": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "occupied": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "last_updated": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-18T09:00:00Z"
                    }
                },
                "type": "object"
            },
            "ParkingSlot": {
                "title": "Parking Slot",
                "description": "Thông tin một chỗ đỗ trong bãi xe",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slot_code": {
                        "type": "string",
                        "example": "B1-001"
                    },
                    "vehicle_type": {
                        "type": "string",
                        "example": "motorbike"
                    },
                    "status": {
                        "type": "string",
                        "example": "available"
                    },
                    "position_x": {
                        "type": "integer",
                        "example": 5
                    },
                    "position_y": {
                        "type": "integer",
                        "example": 8
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-18T09:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-18T09:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}